requirements.txt
streamlit
nltk
requests
python-dotenv
